Fix (most) race conditions in TimeBucketInfo using sync.Once
The previous flag check wasn't safe for concurrency.  I tried
sync.Mutex, but didn't work due to the Go's memory model and ended up
with sync.Once solution.  There are still a few places where it updates
fields when reading it, which is not a very good idea anyway, but leave
it for now since they are minor use cases at the moment.