polygon backfilling improvements
Added query_start parameter to specify where to start backfilling from rather than just the previously written record. Also made the backfilling run in parallel. It is currently spinning 10 goroutines per core on the system, which should be fine since most of the time it will be just waiting on HTTP responses.